Test moved to a separated file for job tests.
